backend/requirements.txt
fastapi
uvicorn
supabase==2.7.0
pyjwt
sqlalchemy
pydantic
python-dotenv
feedparser
beautifulsoup4
requests
langchain
langchain-google-genai
langgraph
qdrant-client
tenacity
